requirements.txt
pytest==7.4.0
requests==2.32.5
pyyaml==6.0
jsonschema==4.19.0
pytest-html==3.2.0
requests-toolbelt==0.10.1
pytest-cov==4.1.0
